feat: rename world_viewer module to "planviewer"
