+++ src/a22_planviewer/planview_server.py
@@ -1,6 +1,6 @@
 from flask import Flask, jsonify, request
 from flask_cors import CORS
-from src.a22_world_viewer.planview_filters import (
+from src.a22_planviewer.planview_filters import (
     plan_awardees,
     plan_facts,
     plan_fund,
